Initialization: Split Files
The `ImpactX.cpp` files got quite long and spaghetti-ish,
so it is time to introduce some structure.

Split out initialization logic, so that we can compile
individual `.cpp` files in parallel.
