+++ src/initialization/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources(ImpactX
+  PRIVATE
+    InitDistribution.cpp
+    InitElement.cpp
+    InitMeshRefinement.cpp
+    InitParser.cpp
+)
