+++ src/initialization/InitParser.cpp
@@ -1,0 +1,23 @@
+/* Copyright 2021 Axel Huebl, Chad Mitchell, Ji Qiang
+ *
+ * This file is part of ImpactX.
+ *
+ * License: BSD-3-Clause-LBNL
+ */
+#include "InitParser.H"
+
+#include <AMReX_ParmParse.H>
+
+namespace impactx
+{
+    void
+    overwrite_amrex_parser_defaults ()
+    {
+        amrex::ParmParse pp_amrex("amrex");
+
+        // https://amrex-codes.github.io/amrex/docs_html/GPU.html#inputs-parameters
+        bool abort_on_out_of_gpu_memory = true; // AMReX' default: false
+        pp_amrex.query("abort_on_out_of_gpu_memory", abort_on_out_of_gpu_memory);
+        pp_amrex.add("abort_on_out_of_gpu_memory", abort_on_out_of_gpu_memory);
+    }
+} // namespace impactx
